scripts for object privileges
find_all_privs_sum.sql - summaries object privileges rather than list them out individually. This is useful for schemas with large amounts of granted privileges.
find_all_privs2.sql -  to find all of the privileges granted to a particular user. It includes Roles, system privileges and object privileges. If a role is encountered then it recursively looks for the roles, system privileges and object privileges granted to the roles and so on.